+++ src/GlobalPayments.Api/Builders/AuthorizationBuilder.cs
@@ -97,6 +97,7 @@ namespace GlobalPayments.Api.Builders {
         internal StoredCredentialInitiator? TransactionInitiator { get; set; }
         internal BNPLShippingMethod BNPLShippingMethod {get;set;}
         internal bool MaskedDataResponse { get; set; }
+        internal MerchantCategory? MerchantCategory { get; set; }
         internal bool HasEmvFallbackData {
             get {
                 return (EmvFallbackCondition != null || EmvLastChipRead != null || !string.IsNullOrEmpty(PaymentApplicationVersion));
@@ -847,9 +848,13 @@ namespace GlobalPayments.Api.Builders {
             return this;
         }
 
-        public AuthorizationBuilder WithMaskedDataResponse(bool value)
-        {
+        public AuthorizationBuilder WithMaskedDataResponse(bool value) {
             MaskedDataResponse = value;
+            return this;
+        }
+
+        public AuthorizationBuilder WithMerchantCategory(MerchantCategory value) {
+            MerchantCategory = value;
             return this;
         }
 
@@ -987,7 +992,27 @@ namespace GlobalPayments.Api.Builders {
 
             Validations.For(TransactionType.Sale)
                 .With(TransactionModifier.EncryptedMobile)
-                .Check(() => PaymentMethod).IsNotNull();            
+                .Check(() => PaymentMethod).IsNotNull();          
+
+            Validations.For(TransactionType.Sale)
+                .With(TransactionModifier.AlternativePaymentMethod)
+                .Check(() => PaymentMethod).IsNotNull()
+                .Check(() => Amount).IsNotNull()
+                .Check(() => Currency).IsNotNull()                
+                .Check(() => PaymentMethod).PropertyOf(nameof(AlternativePaymentMethod.StatusUpdateUrl)).IsNotNull()
+                .Check(() => PaymentMethod).PropertyOf(nameof(AlternativePaymentMethod.ReturnUrl)).IsNotNull()
+                .Check(() => PaymentMethod).PropertyOf(nameof(AlternativePaymentMethod.AccountHolderName)).IsNotNull();
+
+            Validations.For(TransactionType.Auth)
+                .With(TransactionModifier.AlternativePaymentMethod)
+                .Check(() => PaymentMethod).IsNotNull()
+                .Check(() => Amount).IsNotNull()
+                .Check(() => Currency).IsNotNull()
+                .Check(() => PaymentMethod).PropertyOf(nameof(AlternativePaymentMethod.StatusUpdateUrl)).IsNotNull()
+                .Check(() => PaymentMethod).PropertyOf(nameof(AlternativePaymentMethod.ReturnUrl)).IsNotNull()
+                .Check(() => PaymentMethod).PropertyOf(nameof(AlternativePaymentMethod.AccountHolderName)).IsNotNull();
+
+
         }
 
         /// <summary>