+++ src/GlobalPayments.Api/Entities/Enums/FileType.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GlobalPayments.Api.Entities.Enums
+{
+    public enum FileType {
+        TIF,
+        TIFF,
+        BMP,
+        JPG,
+        JPEG,
+        GIF,
+        PNG,
+        DOC,
+        DOCX
+    }
+}