+++ src/GlobalPayments.Api/Entities/Enums/FundsStatus.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GlobalPayments.Api.Entities.Enums
+{
+    public enum FundsStatus {
+        CAPTURED,
+        DECLINE
+    }
+}