+++ src/GlobalPayments.Api/Mapping/GpApiMapping.cs
@@ -25,6 +25,8 @@ namespace GlobalPayments.Api.Mapping {
         private const string ADDRESS_LIST = "ADDRESS_LIST";
         private const string SPLIT = "SPLIT";
         private const string TRANSFER = "TRANSFER";
+        private const string DOCUMENT_UPLOAD = "DOCUMENT_UPLOAD";
+        private const string FUNDS = "FUNDS";
 
         public static Transaction MapResponse(string rawResponse) {
             Transaction transaction = new Transaction();
@@ -175,10 +177,10 @@ namespace GlobalPayments.Api.Mapping {
             return transaction;
         }
 
-        private static List<TransferFundsAccountDetails> MapTransferFundsAccountDetails(JsonDoc json) {
-            var transferResponse = new List<TransferFundsAccountDetails>();
+        private static List<FundsAccountDetails> MapTransferFundsAccountDetails(JsonDoc json) {
+            var transferResponse = new List<FundsAccountDetails>();
             foreach (var item in json.GetArray<JsonDoc>("transfers") ?? Enumerable.Empty<JsonDoc>()) {
-                var transfer = new TransferFundsAccountDetails();
+                var transfer = new FundsAccountDetails();
                 transfer.Id = item.GetValue<string>("id");
                 transfer.Status = item.GetValue<string>("status");
                 transfer.TimeCreated = item.GetValue<string>("time_created");
@@ -231,8 +233,15 @@ namespace GlobalPayments.Api.Mapping {
             JsonDoc json = JsonDoc.Parse(rawResponse);
 
             var paymentMethodApm = json.Get("payment_method")?.Get("apm");
-            apm.RedirectUrl = json.Get("payment_method")?.GetValue<string>("redirect_url");
-            apm.ProviderName = paymentMethodApm?.GetValue<string>("provider");
+            apm.RedirectUrl = json.Get("payment_method")?.GetValue<string>("redirect_url") ?? (paymentMethodApm.GetValue<string>("redirect_url") ?? null);
+            if (paymentMethodApm.Has("provider")) {
+                apm.ProviderName = paymentMethodApm?.Get("provider").GetValue<string>("name") ?? null;
+                apm.ProviderReference = paymentMethodApm?.Get("provider").GetValue<string>("merchant_identifier") ?? null;
+                apm.TimeCreatedReference = paymentMethodApm?.Get("provider").GetValue<DateTime?>("time_created_reference") ?? null;
+            }
+            else {
+                apm.ProviderName = paymentMethodApm?.GetValue<string>("provider");
+            }
             apm.Ack = paymentMethodApm?.GetValue<string>("ack");
             apm.SessionToken = paymentMethodApm?.GetValue<string>("session_token");
             apm.CorrelationReference = paymentMethodApm?.GetValue<string>("correlation_reference");
@@ -1012,6 +1021,24 @@ namespace GlobalPayments.Api.Mapping {
                 JsonDoc json = JsonDoc.Parse(rawResponse);                
                 string actionType = json.Get("action")?.GetValue<string>("type");
                 switch (actionType) {
+                    case DOCUMENT_UPLOAD:
+                        var userDoc = new User();
+                        userDoc.UserReference = new UserReference();
+                        userDoc.UserReference.UserId = json.GetValue<string>("merchant_id") ?? null;
+                        userDoc.UserReference.UserType = UserType.MERCHANT;
+                        userDoc.Name = json.GetValue<string>("merchant_name") ?? null;
+                        var doc = new Document();
+                        doc.Name = json.GetValue<string>("name");
+                        doc.Id = json.GetValue<string>("id");
+                        doc.Status = json.GetValue<string>("status");
+                        doc.TimeCreated = json.GetValue<string>("time_created");
+                        doc.Format = (FileType)Enum.Parse(typeof(FileType),json.GetValue<string>("format"));
+                        doc.Category = (DocumentCategory)Enum.Parse(typeof(DocumentCategory), json.GetValue<string>("function"));
+                        userDoc.Document = doc;
+                        if (json.Has("action")) {
+                            userDoc.ResponseCode = json.Get("action").GetValue<string>("result_code") ?? null;
+                        }                
+                        return userDoc as T;
                     case MERCHANT_CREATE:
                     case MERCHANT_EDIT:
                     case MERCHANT_EDIT_INITIATED:
@@ -1056,6 +1083,25 @@ namespace GlobalPayments.Api.Mapping {
                         }
 
                         return user as T;
+                    case FUNDS:
+                        var userFunds = new User();
+                        userFunds.UserReference = new UserReference();
+                        userFunds.UserReference.UserId = json.GetValue<string>("merchant_id");
+                        userFunds.Name = json.GetValue<string>("merchant_name");
+                        var funds = new FundsAccountDetails();
+                        funds.Id = json.GetValue<string>("id");
+                        funds.TimeCreated = json.GetValue<string>("time_created");
+                        funds.TimeLastUpdated = json.GetValue<string>("time_last_updated") ?? null;
+                        funds.PaymentMethodType = json.GetValue<string>("type") ?? null;
+                        funds.PaymentMethodName = json.GetValue<string>("payment_method") ?? null;
+                        funds.Status = json.GetValue<string>("status");
+                        funds.Amount = json.GetValue<decimal>("amount");
+                        funds.Currency = json.GetValue<string>("currency") ?? null;
+                        funds.Account = new UserAccount(json.GetValue<string>("account_id"), json.GetValue<string>("account_name"));
+                        userFunds.FundsAccountDetails = funds;
+                        userFunds.ResponseCode = json.Get("action").GetValue<string>("result_code");
+
+                        return userFunds as T;
 
                     default:
                         throw new UnsupportedTransactionException("Unknown transaction type " + actionType);