+++ tests/GlobalPayments.Api.Tests/GpApi/BaseGpApiTests.cs
@@ -21,7 +21,7 @@ namespace GlobalPayments.Api.Tests.GpApi {
         protected static readonly string AppKeyForDcc = "Yf6MJDNJKiqObYAb";
         
         protected static readonly string AppIdForBatch = "P3LRVjtGRGxWQQJDE345mSkEh2KfdAyg";
-        protected static readonly string AppKeyForBatch = "ockJr6pv6KFoGiZA";                
+        protected static readonly string AppKeyForBatch = "ockJr6pv6KFoGiZA";
 
         protected static readonly int ExpMonth = DateTime.Now.Month;
         protected static readonly int ExpYear = DateTime.Now.Year + 1;