+++ tests/GlobalPayments.Api.Tests/GpApi/GpApiMerchantAccountsTest.cs
@@ -647,7 +647,171 @@ namespace GlobalPayments.Api.Tests.GpApi
                 Assert.IsTrue(exceptionCaught);
             }
         }
-       
+
+        [TestMethod]
+        public void AddFunds()
+        {
+            var amount = "10";
+            var currency = "USD";
+            var accountId = "FMA_a78b841dfbd14803b3a31e4e0c514c72";
+            var merchantId = "MER_5096d6b88b0b49019c870392bd98ddac";
+            var merchant = User.FromId(merchantId, UserType.MERCHANT);
+        
+            var response = merchant.AddFunds()
+                .WithAmount(amount)
+                .WithAccountNumber(accountId)
+                .WithPaymentMethodName(PaymentMethodName.BankTransfer)
+                .WithPaymentMethodType(PaymentMethodType.Credit)
+                .WithCurrency(currency)
+                .Execute();
+
+            Assert.IsNotNull(response);
+            Assert.AreEqual("SUCCESS", response.ResponseCode);
+            Assert.IsNotNull(response.FundsAccountDetails);
+            Assert.AreEqual(FundsStatus.CAPTURED.ToString(), response.FundsAccountDetails.Status);
+            Assert.AreEqual(amount, response.FundsAccountDetails.Amount.ToString());
+            Assert.AreEqual(currency, response.FundsAccountDetails.Currency);
+            Assert.AreEqual("CREDIT", response.FundsAccountDetails.PaymentMethodType);
+            Assert.AreEqual("BANK_TRANSFER", response.FundsAccountDetails.PaymentMethodName);
+            Assert.IsNotNull(response.FundsAccountDetails.Account);
+            Assert.AreEqual(accountId, response.FundsAccountDetails.Account.Id);
+        }
+
+        [TestMethod]
+        public void AddFunds_OnlyMandatory()
+        {
+            var amount = "10";
+            var accountId = "FMA_a78b841dfbd14803b3a31e4e0c514c72";
+            var merchantId = "MER_5096d6b88b0b49019c870392bd98ddac";
+            var merchant = User.FromId(merchantId, UserType.MERCHANT);
+
+            var response = merchant.AddFunds()
+                .WithAmount(amount)
+                .WithAccountNumber(accountId)                
+                .WithPaymentMethodType(PaymentMethodType.Credit)
+                .Execute();
+
+            Assert.IsNotNull(response);
+            Assert.AreEqual("SUCCESS", response.ResponseCode);
+            Assert.IsNotNull(response.FundsAccountDetails);
+            Assert.AreEqual(FundsStatus.CAPTURED.ToString(), response.FundsAccountDetails.Status);
+            Assert.AreEqual(amount, response.FundsAccountDetails.Amount.ToString());
+            Assert.AreEqual("CREDIT", response.FundsAccountDetails.PaymentMethodType);
+            Assert.AreEqual("BANK_TRANSFER", response.FundsAccountDetails.PaymentMethodName);
+            Assert.IsNotNull(response.FundsAccountDetails.Account);
+            Assert.AreEqual(accountId, response.FundsAccountDetails.Account.Id);
+        }
+
+        [TestMethod]
+        public void AddFunds_InsufficientFunds()
+        {
+            var amount = "10";
+            var accountId = "FMA_a78b841dfbd14803b3a31e4e0c514c72";
+            var merchantId = "MER_5096d6b88b0b49019c870392bd98ddac";
+            var merchant = User.FromId(merchantId, UserType.MERCHANT);
+
+            var response = merchant.AddFunds()
+                .WithAmount(amount)
+                .WithAccountNumber(accountId)                
+                .Execute();
+
+            Assert.IsNotNull(response);
+            Assert.AreEqual("DECLINED", response.ResponseCode);
+            Assert.IsNotNull(response.FundsAccountDetails);
+            Assert.AreEqual(FundsStatus.DECLINE.ToString(), response.FundsAccountDetails.Status);
+            Assert.AreEqual(amount, response.FundsAccountDetails.Amount.ToString());
+            Assert.AreEqual("DEBIT", response.FundsAccountDetails.PaymentMethodType);
+            Assert.AreEqual("BANK_TRANSFER", response.FundsAccountDetails.PaymentMethodName);
+            Assert.IsNotNull(response.FundsAccountDetails.Account);
+            Assert.AreEqual(accountId, response.FundsAccountDetails.Account.Id);
+        }
+
+        [TestMethod]
+        public void AddFunds_WithoutAmount()
+        {
+            var currency = "USD";
+            var accountId = "FMA_a78b841dfbd14803b3a31e4e0c514c72";
+            var merchantId = "MER_5096d6b88b0b49019c870392bd98ddac";
+            var merchant = User.FromId(merchantId, UserType.MERCHANT);
+
+            var errorFound = false;
+            try
+            {
+                var response = merchant.AddFunds()
+                 .WithAccountNumber(accountId)                 
+                 .WithPaymentMethodName(PaymentMethodName.BankTransfer)
+                 .WithPaymentMethodType(PaymentMethodType.Credit)
+                 .WithCurrency(currency)
+                 .Execute();
+            }
+            catch (BuilderException ex)
+            {
+                errorFound = true;
+                Assert.AreEqual("Amount cannot be null for this transaction type.", ex.Message);
+            }
+            finally {
+                Assert.IsTrue(errorFound);
+            }
+        }
+
+        [TestMethod]
+        public void AddFunds_WithoutAccountNumber()
+        {
+            var amount = "10";
+            var currency = "USD";            
+            var merchantId = "MER_5096d6b88b0b49019c870392bd98ddac";
+            var merchant = User.FromId(merchantId, UserType.MERCHANT);
+
+            var errorFound = false;
+            try
+            {
+                var response = merchant.AddFunds()
+                 .WithAmount(amount)                 
+                 .WithPaymentMethodName(PaymentMethodName.BankTransfer)
+                 .WithPaymentMethodType(PaymentMethodType.Credit)
+                 .WithCurrency(currency)
+                 .Execute();
+            }
+            catch (BuilderException ex)
+            {
+                errorFound = true;
+                Assert.AreEqual("AccountNumber cannot be null for this transaction type.", ex.Message);
+            }
+            finally
+            {
+                Assert.IsTrue(errorFound);
+            }
+        }
+
+        [TestMethod]
+        public void AddFunds_WithoutUserRef()
+        {
+            var amount = "10";
+            var currency = "USD";
+            var accountId = "FMA_a78b841dfbd14803b3a31e4e0c514c72";                     
+
+            var errorFound = false;
+            try
+            {
+                var response = new User().AddFunds()
+                 .WithAmount(amount)
+                 .WithAccountNumber(accountId)
+                 .WithPaymentMethodName(PaymentMethodName.BankTransfer)
+                 .WithPaymentMethodType(PaymentMethodType.Credit)
+                 .WithCurrency(currency)
+                 .Execute();
+            }
+            catch (GatewayException ex)
+            {
+                errorFound = true;
+                Assert.AreEqual("property UserId or config MerchantId cannot be null for this transactionType", ex.Message);
+            }
+            finally
+            {
+                Assert.IsTrue(errorFound);
+            }
+        }
+
         #endregion